requirements.txt
streamlit
PyMuPDF
python-docx
sentence-transformers
torch
pytesseract
pillow